refactoring with Abstract COW List class
