+++ warp10/src/main/java/io/warp10/continuum/gts/COWList.java
@@ -16,22 +16,15 @@
 
 package io.warp10.continuum.gts;
 
-import java.util.ArrayList;
 import java.util.BitSet;
-import java.util.Collection;
-import java.util.Comparator;
-import java.util.Iterator;
 import java.util.List;
-import java.util.ListIterator;
-import java.util.NoSuchElementException;
-import java.util.function.UnaryOperator;
 
 /**
  * Copy On Write List implementation backed by an existing array or bitset.
  * Data is duplicated when the list is modified.
  * The original array or bitset is never modified.
  */
-public class COWList implements List {
+public class COWList extends AbstractCOWList {
 
   private final int size;
   private int startIdx; // index of first element of sublist, inclusive
@@ -41,14 +34,6 @@ public class COWList implements List {
   private double[] dataDouble = null;
   private String[] dataString = null;
   private BitSet dataBoolean = null;
-
-  /**
-   * As long as readOnly is true, the List is backed by the original array (view). You can trust dataType.
-   * As soon as user asks for a modification of the list, data are copied in an ArrayList, and readOnly turns false.
-   * When backed by the ArrayList, dataType can be heterogeneous
-   */
-  private boolean readOnly;
-  private ArrayList mutableCopy = null;
 
   public static enum TYPE {
     LONG, DOUBLE, BOOLEAN, STRING
@@ -136,98 +121,12 @@ public class COWList implements List {
     this.readOnly = true;
   }
 
-  public boolean isReadOnly() {
-    return readOnly;
-  }
-
-  private synchronized void initialDeepCopy() {
-      if (readOnly) {
-        mutableCopy = new ArrayList(virtualSize);
-        switch (dataType) {
-          case LONG:
-            for (int i = 0; i < virtualSize; i++) {
-              mutableCopy.add(dataLong[i + startIdx]);
-            }
-            break;
-          case DOUBLE:
-            for (int i = 0; i < virtualSize; i++) {
-              mutableCopy.add(dataDouble[i + startIdx]);
-            }
-            break;
-          case STRING:
-            for (int i = 0; i < virtualSize; i++) {
-              mutableCopy.add(dataString[i + startIdx]);
-            }
-            break;
-          case BOOLEAN:
-            for (int i = 0; i < virtualSize; i++) {
-              mutableCopy.add(dataBoolean.get(i + startIdx));
-            }
-            break;
-        }
-        readOnly = false;
-      }
-    
-  }
-
   private void initialRangeCheck(int startIndex, int length) {
     if (startIndex >= size || startIndex < 0) {
-      throw new IndexOutOfBoundsException(outOfBoundsMsg(startIndex));
+      throw new IndexOutOfBoundsException("Index: " + startIndex + ", Size: " + size);
     }
     if (length < 0 || (startIndex + length) > size) {
       throw new IndexOutOfBoundsException("Start index(" + startIndex + ") + length(" + length + ") is greater than original array size(" + size + "), cannot create sublist");
-    }
-  }
-
-  private void rangeCheck(int index) {
-    if (index < 0 || index >= virtualSize) {
-      throw new IndexOutOfBoundsException(outOfBoundsMsg(index));
-    }
-  }
-
-  private String outOfBoundsMsg(int index) {
-    return "Index: " + index + ", Size: " + size;
-  }
-
-  @Override
-  public int indexOf(Object o) {
-    if (readOnly) {
-      if (o == null) {
-        return -1; // no null object in a primitive array
-      }
-      switch (dataType) {
-        case LONG:
-          for (int i = 0; i < virtualSize; i++) {
-            if (o.equals(dataLong[i + startIdx])) {
-              return i;
-            }
-          }
-          break;
-        case DOUBLE:
-          for (int i = 0; i < virtualSize; i++) {
-            if (o.equals(dataDouble[i + startIdx])) {
-              return i;
-            }
-          }
-          break;
-        case STRING:
-          for (int i = 0; i < virtualSize; i++) {
-            if (o.equals(dataString[i + startIdx])) {
-              return i;
-            }
-          }
-          break;
-        case BOOLEAN:
-          for (int i = 0; i < virtualSize; i++) {
-            if (o.equals(dataBoolean.get(i + startIdx))) {
-              return i;
-            }
-          }
-          break;
-      }
-      return -1;
-    } else {
-      return mutableCopy.indexOf(o);
     }
   }
 
@@ -238,138 +137,6 @@ public class COWList implements List {
     } else {
       return mutableCopy.size();
     }
-  }
-
-  @Override
-  public boolean isEmpty() {
-    if (readOnly) {
-      return 0 == virtualSize;
-    } else {
-      return mutableCopy.isEmpty();
-    }
-  }
-
-  @Override
-  public boolean contains(Object o) {
-    return indexOf(o) >= 0;
-  }
-  
-  @Override
-  public Iterator iterator() {
-    if (readOnly) {
-      return new Iterator() {
-
-        int cursor;       // index of next element to return
-
-        public boolean hasNext() {
-          return cursor != virtualSize;
-        }
-
-        public Object next() {
-          int i = cursor;
-          if (i >= virtualSize) {
-            throw new NoSuchElementException();
-          }
-          Object res = null;
-          switch (dataType) {
-            case LONG:
-              res = dataLong[startIdx + i];
-              break;
-            case DOUBLE:
-              res = dataDouble[startIdx + i];
-              break;
-            case STRING:
-              res = dataString[startIdx + i];
-              break;
-            case BOOLEAN:
-              res = dataBoolean.get(startIdx + i);
-              break;
-          }
-          cursor = i + 1;
-          return res;
-        }
-
-        public void remove() {
-          // not applicable
-        }
-      };
-    } else {
-      return mutableCopy.iterator();
-    }
-  }
-
-  /**
-   * Modification of the underlying array or bitset is forbidden. 
-   * add will allocate a new ArrayList, then alter this ArrayList.
-   */
-  @Override
-  public boolean add(Object o) {
-    initialDeepCopy();
-    return mutableCopy.add(o);
-  }
-
-  /**
-   * Modification of the underlying array or bitset is forbidden. 
-   * remove will allocate a new ArrayList, then alter this ArrayList.
-   */
-  @Override
-  public boolean remove(Object o) {
-    initialDeepCopy();
-    return mutableCopy.remove(o);
-  }
-
-  /**
-   * Modification of the underlying array or bitset is forbidden. 
-   * addAll will allocate a new ArrayList, then alter this ArrayList.
-   */
-  @Override
-  public boolean addAll(Collection c) {
-    initialDeepCopy();
-    return mutableCopy.addAll(c);
-  }
-
-
-  /**
-   * Modification of the underlying array or bitset is forbidden. 
-   * addAll will allocate a new ArrayList, then alter this ArrayList.
-   */
-  @Override
-  public boolean addAll(int index, Collection c) {
-    initialDeepCopy();
-    return mutableCopy.addAll(index, c);
-  }
-
-
-  /**
-   * Modification of the underlying array or bitset is forbidden. 
-   * replaceAll will allocate a new ArrayList, then alter this ArrayList.
-   */
-  @Override
-  public void replaceAll(UnaryOperator operator) {
-    initialDeepCopy();
-    mutableCopy.replaceAll(operator);
-  }
-
-
-  /**
-   * Modification of the underlying array or bitset is forbidden. 
-   * sort will allocate a new ArrayList, then alter this ArrayList.
-   */
-  @Override
-  public void sort(Comparator c) {
-    initialDeepCopy();
-    mutableCopy.sort(c);
-  }
-
-
-  /**
-   * Modification of the underlying array or bitset is forbidden. 
-   * clear will allocate a new ArrayList, then alter this ArrayList.
-   */
-  @Override
-  public synchronized void clear() {
-      mutableCopy = new ArrayList();
-      readOnly = false;
   }
 
   /**
@@ -396,94 +163,6 @@ public class COWList implements List {
         return dataBoolean.get(index + startIdx);
     }
     return null; // impossible
-  }
-
-
-  /**
-   * Modification of the underlying array or bitset is forbidden. 
-   * set will allocate a new ArrayList, then alter this ArrayList.
-   */
-  @Override
-  public Object set(int index, Object element) {
-    initialDeepCopy();
-    return mutableCopy.set(index, element);
-  }
-
-
-  /**
-   * Modification of the underlying array or bitset is forbidden. 
-   * add will allocate a new ArrayList, then alter this ArrayList.
-   */
-  @Override
-  public void add(int index, Object element) {
-    initialDeepCopy();
-    mutableCopy.add(index, element);
-  }
-
-
-  /**
-   * Modification of the underlying array or bitset is forbidden. 
-   * remove will allocate a new ArrayList, then alter this ArrayList.
-   */
-  @Override
-  public Object remove(int index) {
-    initialDeepCopy();
-    return mutableCopy.remove(index);
-  }
-
-  @Override
-  public int lastIndexOf(Object o) {
-    if (readOnly) {
-      if (o == null) {
-        return -1; // no null object in a primitive array
-      }
-      switch (dataType) {
-        case LONG:
-          for (int i = virtualSize - 1; i >= 0; i--) {
-            if (o.equals(dataLong[i + startIdx])) {
-              return i;
-            }
-          }
-          break;
-        case DOUBLE:
-          for (int i = virtualSize - 1; i >= 0; i--) {
-            if (o.equals(dataDouble[i + startIdx])) {
-              return i;
-            }
-          }
-          break;
-        case STRING:
-          for (int i = virtualSize - 1; i >= 0; i--) {
-            if (o.equals(dataString[i + startIdx])) {
-              return i;
-            }
-          }
-          break;
-        case BOOLEAN:
-          for (int i = virtualSize - 1; i >= 0; i--) {
-            if (o.equals(dataBoolean.get(i + startIdx))) {
-              return i;
-            }
-          }
-          break;
-      }
-      return -1;
-    } else {
-      return mutableCopy.lastIndexOf(o);
-    }
-  }
-
-
-  @Override
-  public ListIterator listIterator() {
-    initialDeepCopy();
-    return mutableCopy.listIterator();
-  }
-
-  @Override
-  public ListIterator listIterator(int index) {
-    initialDeepCopy();
-    return mutableCopy.listIterator(index);
   }
 
   /**
@@ -521,71 +200,5 @@ public class COWList implements List {
     } else {
       return mutableCopy.subList(fromIndex, toIndex);
     }
-  }
-
-  @Override
-  public boolean retainAll(Collection c) {
-    initialDeepCopy();
-    return mutableCopy.retainAll(c);
-  }
-
-  @Override
-  public boolean removeAll(Collection c) {
-    initialDeepCopy();
-    return mutableCopy.removeAll(c);
-  }
-
-  @Override
-  public boolean containsAll(Collection c) {
-    for (Object e: c)
-      if (!contains(e)) {
-        return false;
-      }
-    return true;
-  }
-
-  /**
-   * creates a new long/double/String/boolean/Object array from the current subList (deep copy).
-   * if the array provided as {@code a} is not big enough, toArray allocates and returns a new array
-   */
-  @Override
-  public Object[] toArray(Object[] a) {
-    if (readOnly) {
-      Object[] r = a;
-      if (r.length < virtualSize) {
-        r = new Object[virtualSize];
-      }
-      switch (dataType) {
-        case LONG:
-          for (int i = 0; i < virtualSize; i++) {
-            r[i] = (Long) dataLong[i + startIdx];
-          }
-          break;
-        case DOUBLE:
-          for (int i = 0; i < virtualSize; i++) {
-            r[i] = (Double) dataDouble[i + startIdx];
-          }
-          break;
-        case STRING:
-          System.arraycopy(dataString, 0, r, 0, virtualSize);
-          break;
-        case BOOLEAN:
-          for (int i = 0; i < virtualSize; i++) {
-            r[i] = (Boolean) dataBoolean.get(i + startIdx);
-          }
-          break;
-      }
-      return r;
-    } else {
-      return mutableCopy.toArray(a);
-    }
-  }
-
-  /**
-   * creates a new long/double/String/boolean array from the current subList (deep copy).
-   */
-  @Override
-  public Object[] toArray() {
-    return toArray(new Object[virtualSize]);
   }
 }
