+++ warp10/src/main/java/io/warp10/continuum/gts/COWTList.java
@@ -17,13 +17,7 @@
 package io.warp10.continuum.gts;
 
 import java.util.ArrayList;
-import java.util.Collection;
-import java.util.Comparator;
-import java.util.Iterator;
 import java.util.List;
-import java.util.ListIterator;
-import java.util.NoSuchElementException;
-import java.util.function.UnaryOperator;
 
 /**
  * Copy On Write Transversal List
@@ -32,7 +26,7 @@ import java.util.function.UnaryOperator;
  *
  * For instance, it is used by REDUCE and APPLY frameworks
  */
-public class COWTList implements List {
+public class COWTList extends AbstractCOWList {
 
   public static enum TYPE {
     LOCATIONS, ELEVATIONS, VALUES
@@ -68,39 +62,12 @@ public class COWTList implements List {
     return exposeNullValues;
   }
 
-  /**
-   * As long as readOnly is true, the List is backed by the GTS list (view).
-   * As soon as user asks for a modification of the list, data are copied in an ArrayList, and readOnly turns false.
-   */
-  private boolean readOnly = true;
-  private ArrayList mutableCopy = null;
-
-  public boolean isReadOnly() {
-    return readOnly;
-  }
-
-  private synchronized void initialDeepCopy() {
-    if (readOnly) {
-      mutableCopy = new ArrayList(size());
-      for (int i = 0; i < size(); i++) {
-        mutableCopy.add(get(i));
-      }
-      readOnly = false;
-    }
-  }
-
   @Override
   public int size() {
     if (readOnly) {
       return exposeNullValues ? gtsList.size() : gtsList.size() - skippedGTSIndices.size();
     } else {
       return mutableCopy.size();
-    }
-  }
-
-  private void rangeCheck(int index) {
-    if (index < 0 || index >= size()) {
-      throw new IndexOutOfBoundsException("Index: " + index + ", Size: " + size());
     }
   }
 
@@ -217,181 +184,5 @@ public class COWTList implements List {
     } else {
       return mutableCopy.subList(fromIndex, toIndex);
     }
-  }
-
-  // todo(refactoring): since most of the overrides below are the same than for COWList, they both could extend an AbstractCOWList class
-
-  @Override
-  public int indexOf(Object o) {
-    if (readOnly) {
-      for (int i = 0; i < size(); i++) {
-        if (o.equals(get(i))) {
-          return i;
-        }
-      }
-      return -1;
-    } else {
-      return mutableCopy.indexOf(o);
-    }
-  }
-
-  @Override
-  public int lastIndexOf(Object o) {
-    //todo
-    return 0;
-  }
-
-  @Override
-  public boolean isEmpty() {
-    return 0 == size();
-  }
-
-  @Override
-  public boolean contains(Object o) {
-    return indexOf(o) >= 0;
-  }
-
-  @Override
-  public Iterator iterator() {
-    if (readOnly) {
-      return new Iterator() {
-        int cursor = 0;
-
-        @Override
-        public boolean hasNext() {
-          return cursor < size();
-        }
-
-        @Override
-        public Object next() {
-          if (!hasNext()) {
-            throw new NoSuchElementException();
-          }
-          Object res = get(cursor);
-          cursor++;
-          return res;
-        }
-
-        public void remove() {}
-      };
-
-    } else {
-      return mutableCopy.iterator();
-    }
-  }
-
-  @Override
-  public boolean add(Object o) {
-    initialDeepCopy();
-    return mutableCopy.add(o);
-  }
-
-  @Override
-  public boolean remove(Object o) {
-    initialDeepCopy();
-    return mutableCopy.remove(o);
-  }
-
-  @Override
-  public boolean addAll(Collection c) {
-    initialDeepCopy();
-    return mutableCopy.addAll(c);
-  }
-
-  @Override
-  public boolean addAll(int i, Collection c) {
-    initialDeepCopy();
-    return mutableCopy.addAll(i, c);
-  }
-
-  @Override
-  public void replaceAll(UnaryOperator operator) {
-    initialDeepCopy();
-    mutableCopy.replaceAll(operator);
-  }
-
-  @Override
-  public void sort(Comparator c) {
-    initialDeepCopy();
-    mutableCopy.sort(c);
-  }
-
-  @Override
-  public void clear() {
-    mutableCopy = new ArrayList();
-    readOnly = false;
-  }
-
-  @Override
-  public Object set(int i, Object o) {
-    initialDeepCopy();
-    return mutableCopy.set(i, o);
-  }
-
-  @Override
-  public void add(int i, Object o) {
-    initialDeepCopy();
-    mutableCopy.add(i, o);
-  }
-
-  @Override
-  public Object remove(int i) {
-    initialDeepCopy();
-    return mutableCopy.remove(i);
-  }
-
-  @Override
-  public ListIterator listIterator() {
-    initialDeepCopy();
-    return mutableCopy.listIterator();
-  }
-
-  @Override
-  public ListIterator listIterator(int i) {
-
-    initialDeepCopy();
-    return mutableCopy.listIterator(i);
-  }
-
-  @Override
-  public boolean retainAll(Collection c) {
-    initialDeepCopy();
-    return mutableCopy.retainAll(c);
-  }
-
-  @Override
-  public boolean removeAll(Collection c) {
-    initialDeepCopy();
-    return mutableCopy.removeAll(c);
-  }
-
-  @Override
-  public boolean containsAll(Collection c) {
-    for (Object e: c)
-      if (!contains(e)) {
-        return false;
-      }
-    return true;
-  }
-
-  @Override
-  public Object[] toArray(Object[] a) {
-    if (readOnly) {
-      Object[] r = a;
-      if (r.length < size()) {
-        r = new Object[size()];
-      }
-      for (int i = 0; i < size(); i++) {
-        r[i] = get(i);
-      }
-      return r;
-    } else {
-      return mutableCopy.toArray(a);
-    }
-  }
-
-  @Override
-  public Object[] toArray() {
-    return toArray(new Object[size()]);
   }
 }
